+++ app/src/test/java/org/simple/clinic/newentry/PatientEntryUpdateTest.kt
@@ -11,6 +11,7 @@ import org.simple.clinic.MAX_ALLOWED_PATIENT_AGE
 import org.simple.clinic.MIN_ALLOWED_PATIENT_AGE
 import org.simple.clinic.TestData
 import org.simple.clinic.appconfig.Country
+import org.simple.clinic.newentry.country.IndiaInputFieldsProvider
 import org.simple.clinic.newentry.country.InputFields
 import org.simple.clinic.newentry.country.InputFieldsFactory
 import org.simple.clinic.patient.Gender
@@ -44,11 +45,10 @@ class PatientEntryUpdateTest {
   private val updateSpec = UpdateSpec(update)
   private val defaultModel = PatientEntryModel.DEFAULT
 
-  private val inputFieldsFactory = InputFieldsFactory(
+  private val inputFieldsFactory = InputFieldsFactory(IndiaInputFieldsProvider(
       dateTimeFormatter = dateOfBirthFormat,
-      today = localDate
-  )
-  private val india = TestData.country(isoCountryCode = Country.INDIA)
+      today = LocalDate.now(userClock)
+  ))
 
   @Test
   fun `when the user grants reminder consent, update the model`() {
@@ -431,7 +431,7 @@ class PatientEntryUpdateTest {
 
   @Test
   fun `when the input fields are loaded, the UI muse be initialized`() {
-    val inputFields = InputFields(inputFieldsFactory.fieldsFor(india))
+    val inputFields = InputFields(inputFieldsFactory.provideFields())
 
     updateSpec
         .given(defaultModel)